Extract repository type from function that creates it
