+++ src/registry/routes/component.ts
@@ -3,7 +3,8 @@ import _ from 'lodash';
 
 import GetComponentHelper from './helpers/get-component';
 import strings from '../../resources';
-import { Config, Repository } from '../../types';
+import type { Repository } from '../domain/repository';
+import { Config } from '../../types';
 import { Request, RequestHandler, Response } from 'express';
 
 export default function component(
