+++ src/registry/routes/index.ts
@@ -9,7 +9,8 @@ import getComponentsHistory from './helpers/get-components-history';
 import getAvailableDependencies from './helpers/get-available-dependencies';
 import indexView from '../views';
 import urlBuilder = require('../domain/url-builder');
-import { Author, Component, ParsedComponent, Repository } from '../../types';
+import type { Repository } from '../domain/repository';
+import { Author, Component, ParsedComponent } from '../../types';
 import { NextFunction, Request, Response } from 'express';
 import { IncomingHttpHeaders } from 'http';
 import { PackageJson } from 'type-fest';
